Tum classlar - abstract hatali (Product classimiz), baskette hash tipinde degeri eklemedim -

+++ src/Paid.java
@@ -1,5 +1,5 @@
 public interface Paid {
 
-	public double getPrice();
+	public float getPrice();
 
 }
